Extract AnimatedBackground component
