+++ src/pages/home/report/AnimatedBackground/propTypes.js
@@ -1,0 +1,9 @@
+import {windowDimensionsPropTypes} from '../../../../components/withWindowDimensions';
+
+const propTypes = {
+    ...windowDimensionsPropTypes,
+};
+
+const defaultProps = {};
+
+export {propTypes, defaultProps};
