+++ src/pages/home/report/ReportActionItemCreated.js
@@ -3,9 +3,6 @@ import {Pressable, View} from 'react-native';
 import lodashGet from 'lodash/get';
 import {withOnyx} from 'react-native-onyx';
 import PropTypes from 'prop-types';
-import Animated, {
-    useSharedValue, useAnimatedStyle, useAnimatedSensor, SensorType, withSpring,
-} from 'react-native-reanimated';
 import ONYXKEYS from '../../../ONYXKEYS';
 import RoomHeaderAvatars from '../../../components/RoomHeaderAvatars';
 import ReportWelcomeText from '../../../components/ReportWelcomeText';
@@ -15,10 +12,10 @@ import styles from '../../../styles/styles';
 import OfflineWithFeedback from '../../../components/OfflineWithFeedback';
 import * as Report from '../../../libs/actions/Report';
 import reportPropTypes from '../../reportPropTypes';
-import EmptyStateBackgroundImage from '../../../../assets/images/empty-state_background-fade.png';
 import * as StyleUtils from '../../../styles/StyleUtils';
 import withWindowDimensions, {windowDimensionsPropTypes} from '../../../components/withWindowDimensions';
 import compose from '../../../libs/compose';
+import AnimatedBackground from './AnimatedBackground';
 
 const propTypes = {
     /** The id of the report */
@@ -47,26 +44,6 @@ const defaultProps = {
 const ReportActionItemCreated = (props) => {
     const icons = ReportUtils.getIcons(props.report, props.personalDetails, props.policies);
 
-    // Get data from phone rotation sensor and prep other variables for animation
-    const animatedSensor = useAnimatedSensor(SensorType.ROTATION);
-    const offsetX = useSharedValue((-props.windowWidth / 2));
-    const offsetY = useSharedValue(50);
-
-    // Apply data to create style object
-    const animatedStyles = useAnimatedStyle(() => {
-        const {qx, qy} = animatedSensor.sensor.value;
-        if (props.isSmallScreenWidth) {
-            return {
-                transform: [
-                    // The x vs y here seems wrong but is the way to make it feel right to the user
-                    {translateX: withSpring(offsetX.value - (qy * 65))},
-                    {translateY: withSpring(offsetY.value - (qx * 65))},
-                ],
-            };
-        }
-        return {};
-    });
-
     return (
         <OfflineWithFeedback
             pendingAction={lodashGet(props.report, 'pendingFields.addWorkspaceRoom') || lodashGet(props.report, 'pendingFields.createChat')}
@@ -75,11 +52,7 @@ const ReportActionItemCreated = (props) => {
             onClose={() => Report.navigateToConciergeChatAndDeleteReport(props.report.reportID)}
         >
             <View style={StyleUtils.getReportWelcomeContainerStyle(props.isSmallScreenWidth)}>
-                <Animated.Image
-                    pointerEvents="none"
-                    source={EmptyStateBackgroundImage}
-                    style={[StyleUtils.getReportWelcomeBackgroundImageStyle(props.isSmallScreenWidth), animatedStyles]}
-                />
+                <AnimatedBackground />
                 <View
                     accessibilityLabel="Chat welcome message"
                     style={[styles.p5, StyleUtils.getReportWelcomeTopMarginStyle(props.isSmallScreenWidth)]}
